requirements.txt
torch
torchvision
tqdm
scipy
pillow
